refactor(poster): Create generic progressbarslider
