+++ src/handlers/poster-base/components/ProgressBarSlider.tsx
